Alternative color scheme
Colors in p9p seem to be set in several ways:

* Directly calling allocimage(3) with the desired channel and value
* Directly calling allocimage(3) with the desired channel and constant
  defined in include/draw.h
* Access the predeclared black or white members from the Display struct

Most acme color hacks that I found use the first approach, usually
changing only src/cmd/acme/acme.c. As a result, it makes it a bit more
difficult to scale because the colors are not available outside of acme
and it would be hard to propagate changes across other p9p files to work
on a more general p9p color scheme, let alone maintain it.

So as a result I have merged my .Xresources colors into include/draw.h
and declared and initialised them in the Display struct, so I just have
to do display->mycolor to set a color.

Caveats: the price to pay is that channels are hardcoded (RGB24 in that
case) instead of using screen->chan. I can't really think of a situation
in which it would be an issue for me. Other thing to consider is that
now other p9p applications would need to be themed too as they have lost
their original cosmetics.