Add support for tab expansion (because everyone else is stupid and I am forced to follow along)
Tab expansion inserts spaces instead of TAB character. Number of spaces
is dependent upon your current tab stop setting, which can be changed by
running "Tab n".

As of now, it's not possible to turn it on and off during runtime. You
can however see whether it's compiled or not by executing the command
"Tabexpand". The console will show either 1 or 0. This will be taken
care of in a later commit.